+++ packages/react/src/icons/IconZoomText.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import styles from './Icon.module.css';
-import { IconProps } from './Icon.interface';
+import { IconProps, IconSize } from './Icon.interface';
 
 export const IconZoomText = ({
   ariaLabel = 'zoom-text',
@@ -9,7 +9,7 @@ export const IconZoomText = ({
   ariaHidden = true,
   className = '',
   color,
-  size = 's',
+  size = IconSize.Small,
   style = {},
 }: React.SVGProps<SVGSVGElement> & IconProps) => (
   <svg